requirements.txt
alembic==1.0.3
aliyun-python-sdk-core-v3==2.11.0
aliyun-python-sdk-dysmsapi==1.0.0
amqp==2.3.2
aniso8601==4.0.1
asn1crypto==0.24.0
billiard==3.5.0.5
blinker==1.4
celery==4.2.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.1
Flask==0.12
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
flask-paginate==0.5.1
Flask-Pagination==0.0.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.1.0
mns==1.1.4
Pillow==5.3.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.2
PyMySQL==0.9.2
python-dateutil==2.7.5
python-editor==1.0.3
python-memcached==1.59
pytz==2018.7
qiniu==7.2.2
redis==2.10.6
requests==2.20.1
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
